revert to StyledLink in search styles
