+++ plugin-hrm-form/src/components/common/CategoryWithTooltip.tsx
@@ -17,8 +17,7 @@
 import React from 'react';
 import { Tooltip } from '@material-ui/core';
 
-import { CategoryPill, TagMiddleDot } from '../search/styles';
-import { PillText } from '../../styles';
+import { CategoryPill, TagMiddleDot, PillText } from '../../styles';
 
 /**
  * Given a category, truncates it (if necessary) to make it fit (aprox) in the space of 'UNSPECIFIED/OTHER' string
