+++ plugin-hrm-form/src/translations/en-US/flexUI.json
@@ -83,6 +83,8 @@
   "CaseHeader-NoData": "No Data",
   "ReadMore": "see more",
   "ReadLess": "see less",
+  "ReadMore-trail": "... see more",
+  "ReadLess-trail": "... see less",
   "CaseSummary-NoSummaryProvided": "No Summary provided",
   "SectionEntry-Yes": "Yes",
   "SectionEntry-No": "No",
